+++ home.admin/00mainMenu.sh
@@ -95,6 +95,9 @@ fi
 if [ "${sparko}" == "on" ]; then
   OPTIONS+=(SPARKO "Sparko Webwallet")
 fi
+if [ "${spark}" == "on" ]; then
+  OPTIONS+=(SPARK "Spark Wallet")
+fi
 if [ "${ElectRS}" == "on" ]; then
   OPTIONS+=(ELECTRS "Electrum Rust Server")
 fi
@@ -247,6 +250,9 @@ case $CHOICE in
             ;;
         SPARKO)
             /home/admin/config.scripts/cl-plugin.sparko.sh menu mainnet
+            ;;
+        SPARK)
+            /home/admin/config.scripts/cl.spark.sh menu mainnet
             ;;
         LNBITS)
             /home/admin/config.scripts/bonus.lnbits.sh menu