+++ home.admin/00parallelTestnetServices.sh
@@ -14,6 +14,7 @@ if [ ${#tlnd} -eq 0 ]; then tlnd="off"; fi
 if [ ${#tcrtlWebinterface} -eq 0 ]; then tcrtlWebinterface="off"; fi
 if [ ${#tcl} -eq 0 ]; then tcl="off"; fi
 if [ ${#tsparko} -eq 0 ]; then tsparko="off"; fi
+if [ ${#tspark} -eq 0 ]; then tspark="off"; fi
 
 # show select dialog
 echo "run dialog ..."
@@ -24,6 +25,7 @@ OPTIONS+=(r "RTL for LND $CHAIN" ${trtlWebinterface})
 OPTIONS+=(c "C-lightning on $CHAIN" ${tcl})
 OPTIONS+=(t "RTL for CL on $CHAIN" ${tcrtlWebinterface})
 OPTIONS+=(s "Sparko for CL on $CHAIN" ${tsparko})
+OPTIONS+=(m "Spark Wallet fro CL on $CHAIN" ${tspark})
 
 CHOICES=$(dialog --title ' Additional Services ' \
           --checklist ' use spacebar to activate/de-activate ' \
@@ -162,7 +164,27 @@ else
   echo "# Sparko on $CHAIN Setting unchanged."
 fi
 
-
+# tspark process choice
+choice="off"; check=$(echo "${CHOICES}" | grep -c "m")
+if [ ${check} -eq 1 ]; then choice="on"; fi
+if [ "${tspark}" != "${choice}" ]; then
+  echo "# Spark Wallet on $CHAIN Setting changed .."
+  anychange=1
+  /home/admin/config.scripts/cl.spark.sh ${choice} $CHAIN
+  errorOnInstall=$?
+  if [ "${choice}" =  "on" ]; then
+    if [ ${errorOnInstall} -eq 0 ]; then
+      /home/admin/config.scripts/cl.spark.sh menu $CHAIN
+    else
+      l1="# !!! FAIL on Spark Wallet on $CHAIN install !!!"
+      l2="# Try manual install on terminal after reboot with:"
+      l3="/home/admin/config.scripts/cl.spark.sh on $CHAIN"
+      dialog --title 'FAIL' --msgbox "${l1}\n${l2}\n${l3}" 7 65
+    fi
+  fi
+else
+  echo "# Spark Wallet on $CHAIN Setting unchanged."
+fi
 
 if [ ${anychange} -eq 0 ]; then
      dialog --msgbox "NOTHING CHANGED!\nUse Spacebar to check/uncheck services." 8 58