+++ home.admin/00settingsMenuServices.sh
@@ -25,6 +25,7 @@ if [ ${#lit} -eq 0 ]; then lit="off"; fi
 if [ ${#whitepaper} -eq 0 ]; then whitepaper="off"; fi
 if [ ${#chantools} -eq 0 ]; then chantools="off"; fi
 if [ ${#sparko} -eq 0 ]; then sparko="off"; fi
+if [ ${#spark} -eq 0 ]; then spark="off"; fi
 
 # show select dialog
 echo "run dialog ..."
@@ -58,6 +59,7 @@ fi
 if [ "${lightning}" == "cl" ] || [ "${cl}" == "on" ]; then
   OPTIONS+=(c 'C-Lightning RTL Webinterface' ${crtlWebinterface})
   OPTIONS+=(k 'C-Lightning Sparko WebWallet' ${sparko})
+  OPTIONS+=(m 'C-Lightning Spark Wallet' ${spark})
 fi
 
 CHOICES=$(dialog --title ' Additional Mainnet Services ' \
@@ -494,6 +496,28 @@ if [ "${sparko}" != "${choice}" ]; then
   fi
 else
   echo "# Sparko on mainnet Setting unchanged."
+fi
+
+# spark wallet process choice
+choice="off"; check=$(echo "${CHOICES}" | grep -c "m")
+if [ ${check} -eq 1 ]; then choice="on"; fi
+if [ "${spark}" != "${choice}" ]; then
+  echo "# Spark Wallet on mainnet Setting changed .."
+  anychange=1
+  /home/admin/config.scripts/cl.spark.sh ${choice} mainnet
+  errorOnInstall=$?
+  if [ "${choice}" =  "on" ]; then
+    if [ ${errorOnInstall} -eq 0 ]; then
+      /home/admin/config.scripts/cl.spark.sh menu mainnet
+    else
+      l1="# !!! FAIL on Spark Wallet on mainnet install !!!"
+      l2="# Try manual install on terminal after reboot with:"
+      l3="/home/admin/config.scripts/cl.spark.sh on mainnet"
+      dialog --title 'FAIL' --msgbox "${l1}\n${l2}\n${l3}" 7 65
+    fi
+  fi
+else
+  echo "# Spark Wallet on mainnet Setting unchanged."
 fi
 
 if [ ${anychange} -eq 0 ]; then